PROYECTO/src/medirutas_demo/requirements.txt
# Dependencias necesarias para ejecutar Medirutas Demo

Pillow>=9.0.0
